it belongs to yesterday :eyes:

+++ urp project/Assets/_Scripts/Interfaces/IHittable.cs
@@ -4,5 +4,5 @@ using UnityEngine.Events;
 public interface IHittable
 {
     UnityEvent OnGetHit { get; set; }
-    void GetHit(int damage, GameObject damageDealer);
+    void GetHit(float damage, GameObject damageDealer);
 }
